Moved BatchingQueueFactory to kvutils module as well.
